+++ src/lib/components/FormControls/Textarea/Textarea.spec.tsx
@@ -1,6 +1,6 @@
 import { render } from '@testing-library/react';
 import { describe, expect, it } from 'vitest';
-import { Select } from '../Select';
+import { Select } from '../../Select';
 
 describe.concurrent('Components / Form controls / Select', () => {
   describe.concurrent('A11y', () => {
