+++ src/lib/components/Label/Label.spec.tsx
@@ -4,11 +4,11 @@ import { describe, expect, it } from 'vitest';
 import { Button } from '../Button';
 import { Checkbox } from '../Checkbox';
 import { FileInput } from '../FileInput';
-import { Select } from '../FormControls/Select';
 import { Textarea } from '../FormControls/Textarea';
 import { TextInput } from '../FormControls/TextInput';
 import { ToggleSwitch } from '../FormControls/ToggleSwitch';
 import { Radio } from '../Radio';
+import { Select } from '../Select';
 import { Label } from './Label';
 
 describe.concurrent('Components / Label', () => {
