+++ src/lib/components/Select/Select.spec.tsx
@@ -2,7 +2,7 @@ import { render } from '@testing-library/react';
 import { describe, expect, it } from 'vitest';
 import { Select } from './Select';
 
-describe.concurrent('Components / Form controls / Select', () => {
+describe.concurrent('Components / Select', () => {
   describe.concurrent('A11y', () => {
     it('should have role="combobox" by default', () => {
       const select = render(<Select />).getByRole('combobox');
